Remove old query resolvers
