+++ packages/graph/hash_graph/lib/graph/src/shared/mod.rs
@@ -1,4 +1,3 @@
 pub mod identifier;
 pub mod provenance;
-pub mod query;
 pub mod subgraph;
