+++ packages/graph/hash_graph/lib/graph/src/store/postgres/knowledge/entity/mod.rs
@@ -1,5 +1,4 @@
 mod read;
-mod resolve;
 
 use std::{future::Future, pin::Pin};
