+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/data_type.rs
@@ -1,40 +1,43 @@
-mod resolve;
-
 use async_trait::async_trait;
 use error_stack::{IntoReport, Report, Result, ResultExt};
 use futures::{stream, StreamExt, TryStreamExt};
 use tokio_postgres::GenericClient;
-use type_system::{uri::VersionedUri, LinkType};
+use type_system::{uri::VersionedUri, DataType};
 
 use crate::{
-    ontology::{PersistedLinkType, PersistedOntologyMetadata},
+    ontology::{PersistedDataType, PersistedOntologyMetadata},
     provenance::{CreatedById, OwnedById, UpdatedById},
     shared::identifier::GraphElementIdentifier,
     store::{
         crud::Read,
         postgres::{context::PostgresContext, DependencyContext, DependencyContextRef},
-        AsClient, InsertionError, LinkTypeStore, PostgresStore, QueryError, UpdateError,
+        AsClient, DataTypeStore, InsertionError, PostgresStore, QueryError, UpdateError,
     },
     subgraph::{StructuralQuery, Subgraph},
 };
 
 impl<C: AsClient> PostgresStore<C> {
-    /// Internal method to read a [`PersistedLinkType`] into a [`DependencyContext`].
+    /// Internal method to read a [`PersistedDataType`] into a [`DependencyContext`].
     ///
     /// This is used to recursively resolve a type, so the result can be reused.
-    pub(crate) async fn get_link_type_as_dependency<'a>(
+    pub(crate) async fn get_data_type_as_dependency(
         &self,
-        link_type_id: &VersionedUri,
-        context: DependencyContextRef<'a>,
+        data_type_id: &VersionedUri,
+        context: DependencyContextRef<'_>,
     ) -> Result<(), QueryError> {
-        let _unresolved_link_type = context
-            .referenced_link_types
+        let DependencyContextRef {
+            referenced_data_types,
+            graph_resolve_depths,
+            ..
+        } = context;
+
+        let _unresolved_entity_type = referenced_data_types
             .insert_with(
-                link_type_id,
-                Some(context.graph_resolve_depths.link_type_resolve_depth),
+                data_type_id,
+                Some(graph_resolve_depths.data_type_resolve_depth),
                 || async {
-                    Ok(PersistedLinkType::from(
-                        self.read_versioned_ontology_type(link_type_id).await?,
+                    Ok(PersistedDataType::from(
+                        self.read_versioned_ontology_type(data_type_id).await?,
                     ))
                 },
             )
@@ -45,10 +48,10 @@ impl<C: AsClient> PostgresStore<C> {
 }
 
 #[async_trait]
-impl<C: AsClient> LinkTypeStore for PostgresStore<C> {
-    async fn create_link_type(
+impl<C: AsClient> DataTypeStore for PostgresStore<C> {
+    async fn create_data_type(
         &mut self,
-        link_type: LinkType,
+        data_type: DataType,
         owned_by_id: OwnedById,
         created_by_id: CreatedById,
     ) -> Result<PersistedOntologyMetadata, InsertionError> {
@@ -61,7 +64,7 @@ impl<C: AsClient> LinkTypeStore for PostgresStore<C> {
         );
 
         let (_, metadata) = transaction
-            .create(link_type, owned_by_id, created_by_id)
+            .create(data_type, owned_by_id, created_by_id)
             .await?;
 
         transaction
@@ -74,28 +77,28 @@ impl<C: AsClient> LinkTypeStore for PostgresStore<C> {
         Ok(metadata)
     }
 
-    async fn get_link_type<'f: 'q, 'q>(
+    async fn get_data_type<'f: 'q, 'q>(
         &self,
-        query: &'f StructuralQuery<'q, LinkType>,
+        query: &'f StructuralQuery<'q, DataType>,
     ) -> Result<Subgraph, QueryError> {
         let StructuralQuery {
             ref filter,
             graph_resolve_depths,
         } = *query;
 
-        let subgraphs = stream::iter(Read::<PersistedLinkType>::read(self, filter).await?)
-            .then(|link_type| async move {
+        let subgraphs = stream::iter(Read::<PersistedDataType>::read(self, filter).await?)
+            .then(|data_type| async move {
                 let mut dependency_context = DependencyContext::new(graph_resolve_depths);
 
-                let link_type_id = link_type.metadata().identifier().uri().clone();
+                let data_type_id = data_type.metadata().identifier().uri().clone();
                 dependency_context
-                    .referenced_link_types
-                    .insert(&link_type_id, None, link_type);
+                    .referenced_data_types
+                    .insert(&data_type_id, None, data_type);
 
-                self.get_link_type_as_dependency(&link_type_id, dependency_context.as_ref_object())
+                self.get_data_type_as_dependency(&data_type_id, dependency_context.as_ref_object())
                     .await?;
 
-                let root = GraphElementIdentifier::OntologyElementId(link_type_id);
+                let root = GraphElementIdentifier::OntologyElementId(data_type_id);
 
                 Ok::<_, Report<QueryError>>(dependency_context.into_subgraph(vec![root]))
             })
@@ -108,10 +111,10 @@ impl<C: AsClient> LinkTypeStore for PostgresStore<C> {
         Ok(subgraph)
     }
 
-    async fn update_link_type(
+    async fn update_data_type(
         &mut self,
-        link_type: LinkType,
-        updated_by: UpdatedById,
+        data_type: DataType,
+        updated_by_id: UpdatedById,
     ) -> Result<PersistedOntologyMetadata, UpdateError> {
         let transaction = PostgresStore::new(
             self.as_mut_client()
@@ -121,7 +124,7 @@ impl<C: AsClient> LinkTypeStore for PostgresStore<C> {
                 .change_context(UpdateError)?,
         );
 
-        let (_, metadata) = transaction.update(link_type, updated_by).await?;
+        let (_, metadata) = transaction.update(data_type, updated_by_id).await?;
 
         transaction
             .client
