+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/entity_type.rs
@@ -1,89 +1,59 @@
-mod resolve;
-
 use std::{future::Future, pin::Pin};
 
 use async_trait::async_trait;
 use error_stack::{IntoReport, Report, Result, ResultExt};
 use futures::{stream, FutureExt, StreamExt, TryStreamExt};
 use tokio_postgres::GenericClient;
-use type_system::{uri::VersionedUri, PropertyType};
+use type_system::{uri::VersionedUri, EntityType};
 
 use crate::{
-    ontology::{PersistedOntologyMetadata, PersistedPropertyType},
+    ontology::{PersistedEntityType, PersistedOntologyMetadata},
     provenance::{CreatedById, OwnedById, UpdatedById},
     shared::identifier::GraphElementIdentifier,
     store::{
         crud::Read,
         postgres::{context::PostgresContext, DependencyContext, DependencyContextRef},
-        AsClient, InsertionError, PostgresStore, PropertyTypeStore, QueryError, UpdateError,
+        AsClient, EntityTypeStore, InsertionError, PostgresStore, QueryError, UpdateError,
     },
     subgraph::{EdgeKind, GraphResolveDepths, OutwardEdge, StructuralQuery, Subgraph},
 };
 
 impl<C: AsClient> PostgresStore<C> {
-    /// Internal method to read a [`PersistedPropertyType`] into two [`DependencyContext`]s.
+    #[expect(
+        clippy::too_many_lines,
+        reason = "difficult to shrink the number of lines with destructuring and so many \
+                  variables needing to be passed independently"
+    )]
+    /// Internal method to read a [`PersistedEntityType`] into four [`DependencyContext`]s.
     ///
     /// This is used to recursively resolve a type, so the result can be reused.
-    pub(crate) fn get_property_type_as_dependency<'a: 'b, 'b>(
+    pub(crate) fn get_entity_type_as_dependency<'a: 'b, 'b>(
         &'a self,
-        property_type_id: &'b VersionedUri,
+        entity_type_id: &'a VersionedUri,
         mut dependency_context: DependencyContextRef<'b>,
     ) -> Pin<Box<dyn Future<Output = Result<(), QueryError>> + Send + 'b>> {
         async move {
-            let unresolved_property_type = dependency_context
-                .referenced_property_types
+            let unresolved_entity_type = dependency_context
+                .referenced_entity_types
                 .insert_with(
-                    property_type_id,
+                    entity_type_id,
                     Some(
                         dependency_context
                             .graph_resolve_depths
-                            .property_type_resolve_depth,
+                            .entity_type_resolve_depth,
                     ),
                     || async {
-                        Ok(PersistedPropertyType::from(
-                            self.read_versioned_ontology_type(property_type_id).await?,
+                        Ok(PersistedEntityType::from(
+                            self.read_versioned_ontology_type(entity_type_id).await?,
                         ))
                     },
                 )
                 .await?;
 
-            if let Some(property_type) = unresolved_property_type.cloned() {
-                // TODO: Use relation tables
-                //   see https://app.asana.com/0/0/1202884883200942/f
-                for data_type_ref in property_type.inner().data_type_references() {
+            if let Some(entity_type) = unresolved_entity_type.cloned() {
+                for property_type_ref in entity_type.inner().property_type_references() {
                     dependency_context.edges.insert(
-                        GraphElementIdentifier::OntologyElementId(property_type_id.clone()),
-                        OutwardEdge {
-                            edge_kind: EdgeKind::References,
-                            destination: GraphElementIdentifier::OntologyElementId(
-                                data_type_ref.uri().clone(),
-                            ),
-                        },
-                    );
-                    if dependency_context
-                        .graph_resolve_depths
-                        .data_type_resolve_depth
-                        > 0
-                    {
-                        self.get_data_type_as_dependency(
-                            data_type_ref.uri(),
-                            dependency_context.change_depth(GraphResolveDepths {
-                                data_type_resolve_depth: dependency_context
-                                    .graph_resolve_depths
-                                    .data_type_resolve_depth
-                                    - 1,
-                                ..dependency_context.graph_resolve_depths
-                            }),
-                        )
-                        .await?;
-                    }
-                }
-
-                // TODO: Use relation tables
-                //   see https://app.asana.com/0/0/1202884883200942/f
-                for property_type_ref in property_type.inner().property_type_references() {
-                    dependency_context.edges.insert(
-                        GraphElementIdentifier::OntologyElementId(property_type_id.clone()),
+                        GraphElementIdentifier::OntologyElementId(entity_type_id.clone()),
                         OutwardEdge {
                             edge_kind: EdgeKind::References,
                             destination: GraphElementIdentifier::OntologyElementId(
@@ -97,6 +67,8 @@ impl<C: AsClient> PostgresStore<C> {
                         .property_type_resolve_depth
                         > 0
                     {
+                        // TODO: Use relation tables
+                        //   see https://app.asana.com/0/0/1202884883200942/f
                         self.get_property_type_as_dependency(
                             property_type_ref.uri(),
                             dependency_context.change_depth(GraphResolveDepths {
@@ -110,6 +82,67 @@ impl<C: AsClient> PostgresStore<C> {
                         .await?;
                     }
                 }
+
+                // TODO: Use relation tables
+                //   see https://app.asana.com/0/0/1202884883200942/f
+                for (link_type_id, entity_type_ids) in entity_type.inner().link_type_references() {
+                    dependency_context.edges.insert(
+                        GraphElementIdentifier::OntologyElementId(entity_type_id.clone()),
+                        OutwardEdge {
+                            edge_kind: EdgeKind::References,
+                            destination: GraphElementIdentifier::OntologyElementId(
+                                link_type_id.clone(),
+                            ),
+                        },
+                    );
+
+                    if dependency_context
+                        .graph_resolve_depths
+                        .link_type_resolve_depth
+                        > 0
+                    {
+                        self.get_link_type_as_dependency(
+                            link_type_id,
+                            dependency_context.change_depth(GraphResolveDepths {
+                                link_type_resolve_depth: dependency_context
+                                    .graph_resolve_depths
+                                    .link_type_resolve_depth
+                                    - 1,
+                                ..dependency_context.graph_resolve_depths
+                            }),
+                        )
+                        .await?;
+                    }
+                    for referenced_entity_type_id in entity_type_ids {
+                        dependency_context.edges.insert(
+                            GraphElementIdentifier::OntologyElementId(entity_type_id.clone()),
+                            OutwardEdge {
+                                edge_kind: EdgeKind::References,
+                                destination: GraphElementIdentifier::OntologyElementId(
+                                    referenced_entity_type_id.uri().clone(),
+                                ),
+                            },
+                        );
+
+                        if dependency_context
+                            .graph_resolve_depths
+                            .entity_type_resolve_depth
+                            > 0
+                        {
+                            self.get_entity_type_as_dependency(
+                                referenced_entity_type_id.uri(),
+                                dependency_context.change_depth(GraphResolveDepths {
+                                    entity_type_resolve_depth: dependency_context
+                                        .graph_resolve_depths
+                                        .entity_type_resolve_depth
+                                        - 1,
+                                    ..dependency_context.graph_resolve_depths
+                                }),
+                            )
+                            .await?;
+                        }
+                    }
+                }
             }
 
             Ok(())
@@ -119,10 +152,10 @@ impl<C: AsClient> PostgresStore<C> {
 }
 
 #[async_trait]
-impl<C: AsClient> PropertyTypeStore for PostgresStore<C> {
-    async fn create_property_type(
+impl<C: AsClient> EntityTypeStore for PostgresStore<C> {
+    async fn create_entity_type(
         &mut self,
-        property_type: PropertyType,
+        entity_type: EntityType,
         owned_by_id: OwnedById,
         created_by_id: CreatedById,
     ) -> Result<PersistedOntologyMetadata, InsertionError> {
@@ -136,22 +169,22 @@ impl<C: AsClient> PropertyTypeStore for PostgresStore<C> {
 
         // This clone is currently necessary because we extract the references as we insert them.
         // We can only insert them after the type has been created, and so we currently extract them
-        // after as well. See `insert_property_type_references` taking `&property_type`
+        // after as well. See `insert_entity_type_references` taking `&entity_type`
         let (version_id, metadata) = transaction
-            .create(property_type.clone(), owned_by_id, created_by_id)
+            .create(entity_type.clone(), owned_by_id, created_by_id)
             .await?;
 
         transaction
-            .insert_property_type_references(&property_type, version_id)
+            .insert_entity_type_references(&entity_type, version_id)
             .await
             .change_context(InsertionError)
             .attach_printable_lazy(|| {
                 format!(
-                    "could not insert references for property type: {}",
-                    property_type.id()
+                    "could not insert references for entity type: {}",
+                    entity_type.id()
                 )
             })
-            .attach_lazy(|| property_type.clone())?;
+            .attach_lazy(|| entity_type.clone())?;
 
         transaction
             .client
@@ -163,33 +196,33 @@ impl<C: AsClient> PropertyTypeStore for PostgresStore<C> {
         Ok(metadata)
     }
 
-    async fn get_property_type<'f: 'q, 'q>(
+    async fn get_entity_type<'f: 'q, 'q>(
         &self,
-        query: &'f StructuralQuery<'q, PropertyType>,
+        query: &'f StructuralQuery<'q, EntityType>,
     ) -> Result<Subgraph, QueryError> {
         let StructuralQuery {
             ref filter,
             graph_resolve_depths,
         } = *query;
 
-        let subgraphs = stream::iter(Read::<PersistedPropertyType>::read(self, filter).await?)
-            .then(|property_type| async move {
+        let subgraphs = stream::iter(Read::<PersistedEntityType>::read(self, filter).await?)
+            .then(|entity_type| async move {
                 let mut dependency_context = DependencyContext::new(graph_resolve_depths);
 
-                let property_type_id = property_type.metadata().identifier().uri().clone();
-                dependency_context.referenced_property_types.insert(
-                    &property_type_id,
+                let entity_type_id = entity_type.metadata().identifier().uri().clone();
+                dependency_context.referenced_entity_types.insert(
+                    &entity_type_id,
                     None,
-                    property_type,
+                    entity_type,
                 );
 
-                self.get_property_type_as_dependency(
-                    &property_type_id,
+                self.get_entity_type_as_dependency(
+                    &entity_type_id,
                     dependency_context.as_ref_object(),
                 )
                 .await?;
 
-                let root = GraphElementIdentifier::OntologyElementId(property_type_id);
+                let root = GraphElementIdentifier::OntologyElementId(entity_type_id);
 
                 Ok::<_, Report<QueryError>>(dependency_context.into_subgraph(vec![root]))
             })
@@ -202,9 +235,9 @@ impl<C: AsClient> PropertyTypeStore for PostgresStore<C> {
         Ok(subgraph)
     }
 
-    async fn update_property_type(
+    async fn update_entity_type(
         &mut self,
-        property_type: PropertyType,
+        entity_type: EntityType,
         updated_by: UpdatedById,
     ) -> Result<PersistedOntologyMetadata, UpdateError> {
         let transaction = PostgresStore::new(
@@ -217,22 +250,20 @@ impl<C: AsClient> PropertyTypeStore for PostgresStore<C> {
 
         // This clone is currently necessary because we extract the references as we insert them.
         // We can only insert them after the type has been created, and so we currently extract them
-        // after as well. See `insert_property_type_references` taking `&property_type`
-        let (version_id, metadata) = transaction
-            .update(property_type.clone(), updated_by)
-            .await?;
+        // after as well. See `insert_entity_type_references` taking `&entity_type`
+        let (version_id, metadata) = transaction.update(entity_type.clone(), updated_by).await?;
 
         transaction
-            .insert_property_type_references(&property_type, version_id)
+            .insert_entity_type_references(&entity_type, version_id)
             .await
             .change_context(UpdateError)
             .attach_printable_lazy(|| {
                 format!(
-                    "could not insert references for property type: {}",
-                    property_type.id()
+                    "could not insert references for entity type: {}",
+                    entity_type.id()
                 )
             })
-            .attach_lazy(|| property_type.clone())?;
+            .attach_lazy(|| entity_type.clone())?;
 
         transaction
             .client
